got the pointer to MitsubaLoader's instance data
